+++ services/editor/package.json
@@ -87,6 +87,7 @@
     "@types/node": "^15.0.1",
     "@types/ramda": "0.27.40",
     "@types/react": "^17.0.4",
+    "@types/react-calendar": "^3.1.4",
     "@types/react-dom": "^17.0.3",
     "@types/react-router": "^5.1.13",
     "@types/react-router-dom": "^5.1.7",
